2_etl/requirements.txt
pandas
google-cloud-bigquery
python-dotenv
pyarrow
pandas_gbq
